Standalone script to analyse planform data given in YAML format.
Standalone MatLab script to plot and analyze planform data given in YAML format. The code calculate and plot the 1st and 2nd derivatives and some mold parameters. Which parameter to investigate must be defined in the source code. the value for a given parameter at any spanwise position given in the yml file can be requested.